requirements.txt
#langchain
#chainlit
#tabulate
#pandas
chainlit==0.7.700
langchain-experimental==0.0.62
langchain==0.2.7
langchain_community==0.2.7
langchain_core==0.2.18
plotly==5.22.0
pandas==2.2.2
yfinance==0.2.40
langchain-openai==0.1.16
langgraph==0.1.8
pydantic==2.8.2
langchain.tools==0.1.34
statsmodels==0.14.2
matplotlib==3.9.1
python-dotenv==1.0.1
alpaca_trade_api
transformers
googlenews

